merge thomas

+++ Assets/Scripts/StaminaBar.cs
@@ -4,7 +4,7 @@ using UnityEngine.UI;
 public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
-
+    
     public Gradient gradient;
     public Image fill;
     public void SetMaxStamina(int stamina)
@@ -15,7 +15,7 @@ public class StaminaBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetStamina(int stamina)
+    public void SetStamina(float stamina)
     {
         slider.value = stamina;
